✍️  Correction

+++ TP-Lambda-MusicApp/TP-Lambda-MusicApp/Music.cs
@@ -6,10 +6,10 @@ namespace TP_Lambda_MusicApp
 		public string Artist { get; set; }
 		public string Title { get; set; }
 		public string Genre { get; set; }
-		public int Duration { get; set; }
-		public int NumberOfStreams { get; set; }
+		public ulong Duration { get; set; }
+		public ulong NumberOfStreams { get; set; }
 
-		public Music(string Artist, string Title, string Genre, int Duration, int NumberOfStreams)
+		public Music(string Artist, string Title, string Genre, ulong Duration, ulong NumberOfStreams)
 		{
 			this.Artist = Artist;
 			this.Title = Title;
